Split reading of `/proc/pid/updtime` into module
This commit is a mechanical move of the function that reads uptime data
under the procfs/ module. It's intended to keep the code tidy.
